add phadder groups
