--- src/routes/(app)/committees/nollu/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-  import NotImplemented from "$lib/components/NotImplemented.svelte";
-</script>
-
-<NotImplemented />
