+++ src/routes/(app)/committees/nollu/groups/+page.server.ts
@@ -1,40 +1,18 @@
-import {
-  phadderGroupSchema,
-  phadderMandateFilter,
-} from "$lib/nollning/groups/types";
-import apiNames from "$lib/utils/apiNames";
-import { authorize } from "$lib/utils/authorization";
-import DOMPurify from "isomorphic-dompurify";
-import { fail, message, setError, superValidate } from "sveltekit-superforms";
-import { zod4 } from "sveltekit-superforms/adapters";
-import { z } from "zod";
-import type { ExtendedPrisma } from "$lib/server/extendedPrisma";
+import type { PageServerLoad } from "./$types";
+import { committeeActions, committeeLoad } from "../../committee.server";
+import { getYearOrThrowSvelteError } from "$lib/utils/url.server";
 
-const getPhadderMandates = async (
-  prisma: ExtendedPrisma,
-  memberId: string,
-  year: number,
-) =>
-  prisma.mandate.findMany({
-    where: {
-      memberId,
-      ...phadderMandateFilter(year),
-    },
-    orderBy: [
-      {
-        positionId: "asc", // regular phadder comes before uppdrag
-      },
-      {
-        startDate: "asc",
-      },
-    ],
+export const load: PageServerLoad = async ({ locals, url }) => {
+  const { prisma } = locals;
+  const currentYear = new Date().getFullYear();
+  // Allow to see committees from 1982 to the NEXT year
+  const year = getYearOrThrowSvelteError(url, {
+    upperBound: currentYear + 1,
   });
-
-export const load = async ({ locals }) => {
-  const { user, prisma } = locals;
-  authorize(apiNames.NOLLNING.MANAGE_PHADDER_GROUPS, user);
-
-  const phadderGroups = await prisma.phadderGroup.findMany({
+  const phadderGroups = prisma.phadderGroup.findMany({
+    where: {
+      year,
+    },
     include: {
       nollor: true,
       phaddrar: {
@@ -47,197 +25,10 @@ export const load = async ({ locals }) => {
       createdAt: "asc",
     },
   });
-
-  return {
-    groups: await Promise.all(
-      phadderGroups.map(async (group) => ({
-        ...group,
-        form: await superValidate(group, zod4(phadderGroupSchema)),
-      })),
-    ),
-    form: await superValidate(zod4(createPhadderGroupSchema)),
-  };
+  return committeeLoad(prisma, "nollu", url).then(async (data) => ({
+    ...data,
+    phadderGroups: await phadderGroups,
+  }));
 };
 
-const createPhadderGroupSchema = phadderGroupSchema.omit({
-  id: true,
-});
-const updatePhadderGroupSchema = phadderGroupSchema;
-const deletePhadderGroupSchema = phadderGroupSchema.pick({
-  id: true,
-});
-
-const personSchema = z.object({
-  memberId: z.string().uuid(),
-  groupId: z.string().uuid(),
-});
-
-export const actions = {
-  create: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(createPhadderGroupSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    form.data.description = form.data.description
-      ? DOMPurify.sanitize(form.data.description)
-      : form.data.description;
-    await prisma.phadderGroup.create({
-      data: form.data,
-    });
-    return message(form, {
-      message: "Phaddergruppen skapades",
-      type: "success",
-    });
-  },
-  update: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(updatePhadderGroupSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    form.data.description = form.data.description
-      ? DOMPurify.sanitize(form.data.description)
-      : form.data.description;
-    const res = await prisma.phadderGroup.update({
-      where: {
-        id: form.data.id,
-      },
-      data: form.data,
-    });
-    console.log(res);
-    return message(form, {
-      message: "Phaddergruppen uppdaterad",
-      type: "success",
-    });
-  },
-  delete: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(deletePhadderGroupSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    await prisma.phadderGroup.delete({
-      where: {
-        id: form.data.id,
-      },
-    });
-    return message(form, {
-      message: "Phaddergruppen borttagen",
-      type: "success",
-    });
-  },
-  addNolla: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(personSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    await prisma.phadderGroup.update({
-      where: {
-        id: form.data.groupId,
-      },
-      data: {
-        nollor: {
-          connect: {
-            id: form.data.memberId,
-          },
-        },
-      },
-    });
-    return message(form, {
-      message: "Nolla tillagd",
-      type: "success",
-    });
-  },
-  removeNolla: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(personSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    await prisma.phadderGroup.update({
-      where: {
-        id: form.data.groupId,
-      },
-      data: {
-        nollor: {
-          disconnect: {
-            id: form.data.memberId,
-          },
-        },
-      },
-    });
-    return message(form, {
-      message: "Nolla borttagen",
-      type: "success",
-    });
-  },
-  addPhadder: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(personSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    const group = await prisma.phadderGroup.findUnique({
-      where: {
-        id: form.data.groupId,
-      },
-    });
-    if (!group) return setError(form, "groupId", "Group not found");
-    const mandate = await getPhadderMandates(
-      prisma,
-      form.data.memberId,
-      group.year,
-    ).then((mandates) => mandates?.[0]); // get first
-
-    if (!mandate)
-      return setError(
-        form,
-        "memberId",
-        "Personen hittas inte som phadder det året",
-      );
-    await prisma.phadderGroup.update({
-      where: {
-        id: form.data.groupId,
-      },
-      data: {
-        phaddrar: {
-          connect: {
-            id: mandate.id,
-          },
-        },
-      },
-    });
-    return message(form, {
-      message: "Phadder tillagd",
-      type: "success",
-    });
-  },
-  removePhadder: async ({ locals, request }) => {
-    const form = await superValidate(request, zod4(personSchema));
-    if (!form.valid) return fail(400, { form });
-    const { prisma } = locals;
-    const group = await prisma.phadderGroup.findUnique({
-      where: {
-        id: form.data.groupId,
-      },
-    });
-    if (!group) return setError(form, "groupId", "Group not found");
-    const mandates = await getPhadderMandates(
-      prisma,
-      form.data.memberId,
-      group?.year,
-    );
-    if (mandates.length === 0)
-      return setError(
-        form,
-        "memberId",
-        "Personen hittas inte som phadder det året",
-      );
-    await prisma.phadderGroup.update({
-      where: {
-        id: form.data.groupId,
-      },
-      data: {
-        phaddrar: {
-          disconnect: mandates.map((m) => ({
-            id: m.id,
-          })),
-        },
-      },
-    });
-    return message(form, {
-      message: "Phadder borttagen",
-      type: "success",
-    });
-  },
-};
+export const actions = committeeActions("nollu");
